renaming `Managers` package to `managers`
