+++ app/src/androidTest/java/ca/klapstein/baudit/managers/BauditPreferencesManagerTest.java
@@ -1,4 +1,4 @@
-package ca.klapstein.baudit.Managers;
+package ca.klapstein.baudit.managers;
 
 import android.support.test.runner.AndroidJUnit4;
 import org.junit.After;
@@ -7,13 +7,13 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 @RunWith(AndroidJUnit4.class)
-public class BauditRemoteManagerTest {
+public class BauditPreferencesManagerTest {
 
-    private BauditRemoteManager bauditRemoteManager;
+    private BauditPreferencesManager bauditPreferencesManager;
 
     @Before
     public void setUp() {
-        bauditRemoteManager = new BauditRemoteManager();
+        bauditPreferencesManager = new BauditPreferencesManager();
     }
 
     @After
